Fixed issue with timezone offsets being incorrectly calculated. Only affected offsets which weren't whole hours, e.g. Tehran (UTC + 3.30).
Zone designators were previously added to overcome issues with how the timestamp could appear in the past or future on Learning Locker when using Date.toISOString(), but this may no longer be an issue. It is probably best practice to include timezones for ambiguity anyway.

Possibly look to replace the Utils moment.js.